mcp_servers/servers/feedmob/pyproject.toml
[project]
name = "feedmob"
version = "0.1.0"
description = "Feedmob MCP server"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
 "mcp>=1.3.0",
 "psycopg2-binary>=2.9.9",
 "sqlalchemy>=2.0.0",
 "python-dotenv>=1.0.0",
]
[[project.authors]]
name = "Jason"
email = "jason@feedmob.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/feedmob"]

[project.scripts]
feedmob = "feedmob:main"
